LPS-86602 Move integration test class to Test module
